requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database
duckdb>=0.9.0

# Streaming (optional - needs Kafka running)
kafka-python>=2.0.2

# Orchestration (optional - needs Airflow setup)
apache-airflow>=2.7.0

# Big Data (optional - needs JVM)
pyspark>=3.4.0

# Utilities
python-dateutil>=2.8.2
